Change to double because not arduino anymore

+++ Lolin-Pointer/src/AstroClock.h
@@ -11,26 +11,26 @@
 
 class AstroClock {
 private:
-	float lat;
-	float lon;
-	float convertJ2000(uint32_t timestamp);
-	float getLSTd(float j2k);
+	double lat;
+	double lon;
+	double convertJ2000(uint32_t timestamp);
+	double getLSTd(double j2k);
 public:
-	AstroClock(float _lat, float _lon);
+	AstroClock(double _lat, double _lon);
 	virtual ~AstroClock();
 
-	float getLat();
-	void setLat(float _lat);
-	float getLon();
-	void setLon(float _lon);
+	double getLat();
+	void setLat(double _lat);
+	double getLon();
+	void setLon(double _lon);
 
-	float localSiderealTime(uint32_t timestamp);
+	double localSiderealTime(uint32_t timestamp);
 
-	void convert(uint32_t timestamp, float ra, float dec, float* alt,
-			float* az);
+	void convert(uint32_t timestamp, double ra, double dec, double* alt,
+			double* az);
 
-	void unconvert(uint32_t timestamp, float alt, float az, float* ra,
-			float* dec);
+	void unconvert(uint32_t timestamp, double alt, double az, double* ra,
+			double* dec);
 };
 
 #endif /* ASTROCLOCK_H_ */
